pyproject.toml
[tool.poetry]
name = "es-tests"
version = "0.1.0"
description = ""
authors = ["gsantoro <giuseppe.santoro@elastic.co>"]
readme = "README.md"
packages = [{include = "es_tests"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
jinja2 = "^3.1.2"
pytest = "^7.3.1"
structlog = "^23.1.0"
pandas = "^2.0.1"
tabulate = "^0.9.0"
requests-to-curl = "^1.1.0"
pyyaml = "^6.0"
typer = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
